Map gen (rought) finalised
Map generates randomly, a few times overlapping occures and aome doors get blocked - this only happens 1 in every 5 or 6 -.

The rooms are a tad bit wonky but shouldnt be to hard to fix at a later point.

exit room defined however no propper way to tell if this works yet.