Enemy AI roughly done
Enemys will shoot at player, but not damge them, they will chase the player and move around in an idle area.

however there is currently no way to automatically switch between the states, but i have a few ideas i can try to impliment later.

The enemy should also lock on to the targert and face them when following and shooting.